requirements.txt
redis
pandas
yfinance
psycopg2-binary
python-dotenv
schedule
